backend/requirements.txt
croniter ~= 6.0.0
Django ~= 5.2
djangorestframework ~= 3.16.0
django-environ ~= 0.12.0
django-q2 ~= 1.7.6
httpx ~= 0.28.1
psycopg[binary] ~= 3.2.6